requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-ulid==1.1.0
PyYAML==6.0.2
redis==5.3.0
redis-om==0.3.5
rich==14.0.0
rich-toolkit==0.14.7
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.16.0
types-cffi==1.17.0.20250523
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250529
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
